new git 1399-6-3 Edit SahamSalanehController & SahamSalanehService

+++ src/main/java/com/mahdi/sandogh/model/sahamsalaneh/SahamSalaneh.java
@@ -2,10 +2,14 @@ package com.mahdi.sandogh.model.sahamsalaneh;
 
 
 import com.mahdi.sandogh.model.audit.DateAudit;
+import com.mahdi.sandogh.model.bank.Bank;
+import com.mahdi.sandogh.model.fund.Fund;
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.util.HashSet;
+import java.util.Set;
 
 @Setter
 @Getter
@@ -14,12 +18,13 @@ import javax.persistence.*;
 public class SahamSalaneh extends DateAudit { // سهام سالانه
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "yesrs")
-    private int yesrs; //سال
+    private Integer yesrs; //سال
 
     @Column(name = "membershipfee")
-    private long membershipFee; // حق عضویت
+    private Long membershipFee; // حق عضویت
+
 
 }
